fix(pipelines): address CodeRabbit PR #281 review (1 critical + 10 major + 11 minor)
CRITICAL
- internal/api/router_routes.go: /api/v1/internal/pipelines/save now uses
  internalAuth instead of authed. The previous wrap-with-authed made the
  route unreachable for the sidecar (no JWT) AND callable by any logged-in
  user (sidecar identity-claim trust without X-Internal-Token check).
  Route moved into the internalAuth block where the rest of /internal/*
  lives. Sidecar's proxyIPCJSON sets only X-Internal-Token, so this
  closes the sidecar-401 path AND the user-can-impersonate-sidecar path.

MAJOR
- pipeline-run-node.tsx: role=button div is now keyboard-accessible —
  added tabIndex=0, Enter/Space keydown handler, focus ring.
- use-pipelines.ts: AbortController guards against stale fetches across
  workspace/slug switches. Same pattern in usePipelines + usePipelineRuns.
- pipelines.go ListRuns: removed dead 7-column "fast path" (the scanner
  could never decode it). Now a single json_extract query — modernc.org
  /sqlite supports it. Added rows.Err() check.
- pipelines.go InternalSave: now runs full semantic validation —
  agent_slug cross-reference (against author crew's agents),
  pipeline slug cross-reference, and CycleDetect over the workspace
  registry — before persisting. Cross-crew/loop bugs caught at save
  time, not at first run.
- executor.go runCallPipelineStep: nested inputs are now rendered
  against the parent's render context before dispatch. String
  values flow through Render so {{ inputs.X }} / {{ steps.Y.output }}
  resolves the way it does for the prompt itself.
- executor.go runCallPipelineStep: depth+1 propagated through to
  the recursive runDSL call so MaxNestedPipelineDepth fires for
  long chains. Save-time CycleDetect catches loops; this catches
  long legitimate chains.
- workflow-graph.tsx: empty state guard now allows the canvas to
  render when missions=0 BUT pipelines>0. Fresh workspaces with a
  saved pipeline see the registry row instead of "no missions yet".
- types.go InputSpec: Min/Max changed from *int to *float64 so they
  cover both "type: integer" and "type: number" inputs without
  silently truncating fractional caps.
- store.go: test-run gate also rejects FUTURE timestamps. Without
  this, claiming last_test_run_at = year 9999 always looked fresh
  because time.Since returns negative. 1-minute clock-skew tolerance.
- store_test.go / tier_test.go / pipelines_test.go: in-memory SQLite
  test DB pinned to a single connection (SetMaxOpenConns(1) +
  SetMaxIdleConns(1)). Without this, schemata written on one
  connection are invisible on another — silent test flakiness.

MINOR
- agent_config_resolver.go: lookupCrewNamesForWorkspace now checks
  rows.Err() so a context cancellation mid-iteration is logged
  rather than silently producing a partial map.
- pipelines.go DryRun: malformed JSON bodies now return 400 (was
  silently coerced to empty inputs, masking caller bugs).
- executor_test.go: surface RunDefinition errors via t.Fatalf
  instead of `_` discarding — vacuous res.Status assertions can't
  hide unexpected executor errors.
- journal.go truncateForPreview + system_prompt.go oneLine:
  UTF-8-safe truncation. Walk back to a rune boundary before
  slicing so multi-byte characters at the cap (CJK, emoji) don't
  produce invalid UTF-8 in journal payloads.
- tier_test.go: surface Resolve errors so coverage improvement
  doesn't bury real failures.
- cmd_pipeline.go: defer resp.Body.Close() on every HTTP call.
  Handle json.Marshal errors in the save flow. Guard the hash
  preview slice against unexpectedly short hashes (no more panic
  on `saved.DefinitionHash[:12]` if the server ever returns a
  short hash).

Tests: 50 pipeline tests + 8 API smoke tests still pass. No new
tests added — all fixes either change existing assertions
(error checks) or close infra gaps (close/abort/scoping).

Out of scope of this commit: types.go SaveInput.LastTestRunPassed
restructuring (CodeRabbit #229) — the store's freshness check
already validates server-side, with future-timestamp guard added
above. Restructuring SaveInput shape would break the sidecar IPC
contract; left for a separate refactor PR.

+++ internal/pipeline/types.go
@@ -44,14 +44,21 @@ type ExecutionTier struct {
 // InputSpec declares one named input the pipeline accepts. Type matches
 // JSON Schema primitive types so the validation layer can reuse the
 // same semantics for inputs and step outputs.
+//
+// Min/Max are float64 so they cover both `type: integer` and
+// `type: number` inputs. JSON numbers don't distinguish ints from
+// floats on the wire, so the validator coerces — typing the bounds
+// as *int would silently truncate fractional caps for number inputs
+// (e.g. Max=0.5 would round to 0). Validation rejects fractional
+// bounds when the input Type is "integer".
 type InputSpec struct {
-	Name        string `json:"name"`
-	Type        string `json:"type"` // string | integer | number | boolean | array | object
-	Required    bool   `json:"required,omitempty"`
-	Default     any    `json:"default,omitempty"`
-	Description string `json:"description,omitempty"`
-	Min         *int   `json:"min,omitempty"`
-	Max         *int   `json:"max,omitempty"`
+	Name        string   `json:"name"`
+	Type        string   `json:"type"` // string | integer | number | boolean | array | object
+	Required    bool     `json:"required,omitempty"`
+	Default     any      `json:"default,omitempty"`
+	Description string   `json:"description,omitempty"`
+	Min         *float64 `json:"min,omitempty"`
+	Max         *float64 `json:"max,omitempty"`
 }
 
 // OutputSpec declares one named output the pipeline produces. Outputs
@@ -222,6 +229,14 @@ type SaveInput struct {
 	// handler has confirmed a test_run within the gate window. The
 	// store does NOT enforce the freshness rule itself — the handler
 	// does, because the gate is policy, not persistence.
+	// LastTestRunAt + LastTestRunPassed encode "the caller already
+	// ran the test_run gate and these are its results". The store's
+	// Save method does NOT trust these naively: it validates the
+	// timestamp is recent (within testRunFreshness) AND not in the
+	// future, both against time.Now() server-side. Callers cannot
+	// mint a passing gate by claiming a fake distant-future
+	// timestamp — the server's clock is the source of truth for
+	// freshness.
 	LastTestRunAt     *time.Time
 	LastTestRunPassed bool
 	// ExecutionTierJSON optional override of workspace tier mapping.
